backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.20
pydantic==2.4.2
Pillow==10.2.0
ultralytics==8.1.2
numpy==1.26.4 